Change package type to 'project'
The 'wordpress-theme' type seemed useless.

Even if you are using a composer-managed WordPress installation,
you are still required to install this theme as a standalone
project that manages its own dependencies anyway, rather than
managed by WordPress.

Will keep a watch on this change though, and revert if it becomes
necessary.